Refactor DataSourceChecker as SPI
