+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-core/src/main/java/org/apache/shardingsphere/scaling/core/job/check/EnvironmentChecker.java
@@ -18,20 +18,12 @@
 package org.apache.shardingsphere.scaling.core.job.check;
 
 import org.apache.shardingsphere.data.pipeline.core.prepare.datasource.DataSourcePreparer;
-import org.apache.shardingsphere.data.pipeline.spi.check.datasource.DataSourceChecker;
 
 /**
  * Environment checker.
  */
 // TODO separate all methods to SPI
 public interface EnvironmentChecker {
-    
-    /**
-     * Get data source checker type.
-     *
-     * @return data source checker type.
-     */
-    Class<? extends DataSourceChecker> getDataSourceCheckerClass();
     
     /**
      * Get data source preparer type.
