+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-core/src/main/java/org/apache/shardingsphere/scaling/core/job/check/EnvironmentCheckerFactory.java
@@ -22,7 +22,6 @@ import org.apache.shardingsphere.data.pipeline.api.config.rulealtered.JobConfigu
 import org.apache.shardingsphere.data.pipeline.core.check.consistency.DataConsistencyChecker;
 import org.apache.shardingsphere.data.pipeline.core.check.consistency.DataConsistencyCheckerImpl;
 import org.apache.shardingsphere.data.pipeline.core.prepare.datasource.DataSourcePreparer;
-import org.apache.shardingsphere.data.pipeline.spi.check.datasource.DataSourceChecker;
 import org.apache.shardingsphere.scaling.core.spi.ScalingEntry;
 import org.apache.shardingsphere.scaling.core.spi.ScalingEntryLoader;
 
@@ -39,18 +38,6 @@ public final class EnvironmentCheckerFactory {
      */
     public static DataConsistencyChecker newInstance(final JobConfiguration jobConfig) {
         return new DataConsistencyCheckerImpl(jobConfig);
-    }
-    
-    /**
-     * Create data source checker instance.
-     *
-     * @param databaseType database type
-     * @return data source checker
-     */
-    @SneakyThrows(ReflectiveOperationException.class)
-    public static DataSourceChecker newInstance(final String databaseType) {
-        ScalingEntry scalingEntry = ScalingEntryLoader.getInstance(databaseType);
-        return scalingEntry.getEnvironmentCheckerClass().getConstructor().newInstance().getDataSourceCheckerClass().getConstructor().newInstance();
     }
     
     /**
