+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-dialect/shardingsphere-data-pipeline-mysql/src/main/java/org/apache/shardingsphere/data/pipeline/mysql/MySQLEnvironmentChecker.java
@@ -18,7 +18,6 @@
 package org.apache.shardingsphere.data.pipeline.mysql;
 
 import org.apache.shardingsphere.data.pipeline.core.prepare.datasource.DataSourcePreparer;
-import org.apache.shardingsphere.data.pipeline.mysql.check.datasource.MySQLDataSourceChecker;
 import org.apache.shardingsphere.data.pipeline.mysql.prepare.datasource.MySQLDataSourcePreparer;
 import org.apache.shardingsphere.scaling.core.job.check.EnvironmentChecker;
 
@@ -26,11 +25,6 @@ import org.apache.shardingsphere.scaling.core.job.check.EnvironmentChecker;
  * MySQL environment checker.
  */
 public final class MySQLEnvironmentChecker implements EnvironmentChecker {
-    
-    @Override
-    public Class<MySQLDataSourceChecker> getDataSourceCheckerClass() {
-        return MySQLDataSourceChecker.class;
-    }
     
     @Override
     public Class<? extends DataSourcePreparer> getDataSourcePreparerClass() {
