+++ shardingsphere-test/shardingsphere-pipeline-test/src/test/java/org/apache/shardingsphere/data/pipeline/core/fixture/FixtureEnvironmentChecker.java
@@ -18,15 +18,9 @@
 package org.apache.shardingsphere.data.pipeline.core.fixture;
 
 import org.apache.shardingsphere.data.pipeline.core.prepare.datasource.DataSourcePreparer;
-import org.apache.shardingsphere.data.pipeline.spi.check.datasource.DataSourceChecker;
 import org.apache.shardingsphere.scaling.core.job.check.EnvironmentChecker;
 
 public final class FixtureEnvironmentChecker implements EnvironmentChecker {
-    
-    @Override
-    public Class<? extends DataSourceChecker> getDataSourceCheckerClass() {
-        return FixtureDataSourceChecker.class;
-    }
     
     @Override
     public Class<? extends DataSourcePreparer> getDataSourcePreparerClass() {
